refactor: refactor project structure of expert web
