--- frontend-web-expert/src/component/header/index.tsx
@@ -1,15 +1,0 @@
-import { Description, OuterDiv, Title } from "./style";
-
-type HeaderProps = {
-  title: string;
-  description: string;
-};
-
-export default function Header({ title, description }: HeaderProps) {
-  return (
-    <OuterDiv>
-      <Title>{title}</Title>
-      <Description>{description}</Description>
-    </OuterDiv>
-  );
-}
