+++ frontend-web-expert/src/component/profile/style.ts
@@ -1,29 +1,6 @@
 import { theme } from "@/style/theme";
-import { keyframes, styled } from "styled-components";
-import { Link, NavLink } from "react-router-dom";
-import { IconButton } from "@mui/material";
-
-const swing = keyframes`
-  0%,
-  30%,
-  50%,
-  70%,
-  100% {
-    transform: rotate(0deg);
-  }
-  10% {
-    transform: rotate(10deg);
-  }
-  40% {
-    transform: rotate(-10deg);
-  }
-  60% {
-    transform: rotate(5deg);
-  }
-  80% {
-    transform: rotate(-5deg);
-  }
-`;
+import { Link } from "react-router-dom";
+import styled, { keyframes } from "styled-components";
 
 export const rotate = keyframes`
   100% {
@@ -38,120 +15,6 @@ export const back = keyframes`
   100% {
     transform: rotateZ(360deg);
   }
-`;
-
-export const Screen = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  min-height: 100dvh;
-`;
-
-export const Header = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  height: 6.75rem;
-  line-height: 6.75rem;
-  box-sizing: border-box;
-`;
-
-export const ButtonBox = styled.div`
-  display: flex;
-  flex-direction: row;
-  padding: 0 1rem;
-`;
-
-export const SwingIconButton = styled(IconButton)`
-  &:hover {
-    color: ${theme.colors.indigo[500]};
-  }
-  &:hover svg {
-    animation: ${swing} ease-in-out 0.5s 1 alternate;
-  }
-`;
-
-export const NonTopSection = styled.div`
-  display: flex;
-  flex-direction: row;
-  height: 100%;
-  background-color: #f5f5f9;
-`;
-
-export const Sidebar = styled.div`
-  & {
-    position: relative;
-    width: 20rem;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-  }
-  &:after {
-    content: "";
-    position: absolute;
-    top: 2rem;
-    bottom: 2rem;
-    right: 0;
-    width: 1px;
-    border-right: 1px solid ${theme.colors.gray[300]};
-  }
-  & > * {
-    padding-left: 1rem;
-    padding-right: 1rem;
-  }
-`;
-
-export const Nav = styled.nav`
-  flex: 1;
-  overflow: auto;
-  padding: 1.6rem 1rem;
-`;
-
-export const Ul = styled.ul`
-  padding: 0;
-  margin: 0;
-  list-style: none;
-`;
-
-export const Item = styled(NavLink)`
-  & {
-    display: flex;
-    align-items: center;
-    overflow: hidden;
-
-    white-space: pre;
-    padding: 0.5rem 5rem 0.5rem 1.5rem;
-    margin: 0.4rem 0;
-    border-radius: 4px;
-    color: inherit;
-
-    text-decoration: none;
-    font-size: 1.05rem;
-    line-height: 2.5;
-  }
-  &:hover {
-    background: ${theme.colors.indigo[200]};
-  }
-  &.active {
-    background: ${theme.colors.indigo[500]};
-    color: white;
-  }
-  &.pending {
-    color: ${theme.colors.gray[200]};
-  }
-`;
-
-export const ItemName = styled.span`
-  margin-left: 1rem;
-`;
-
-export const TopRight = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  margin: 0;
-  padding: 0 2rem;
 `;
 
 export const ProfileImg = styled.img`
@@ -174,8 +37,8 @@ export const ProfileText = styled.div`
   &:after {
     content: "";
     position: absolute;
-    top: 1.2rem;
-    bottom: 1.2rem;
+    top: 2rem;
+    bottom: 2rem;
     left: 0;
     width: 1px;
     border-left: 1px solid ${theme.colors.gray[300]};
@@ -257,8 +120,9 @@ export const ProfileMenuItem = styled(Link)`
     text-decoration: none;
     color: black;
     border-radius: 0.2rem;
-    padding: 0.4rem 0.4rem 0.2rem 0.6rem;
+    padding: 0.5rem;
     margin: 0.1rem 0;
+    line-height: 1;
   }
   &:hover {
     background-color: ${theme.colors.gray[200]};
@@ -292,20 +156,5 @@ export const Logout = styled.button`
   }
   &:hover {
     background-color: ${theme.colors.mojo[700]};
-  }
-`;
-
-export const Detail = styled.div`
-  & {
-    flex: 1;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    padding: 2rem;
-  }
-  &.loading {
-    opacity: 0.25;
-    transition: opacity 200ms;
-    transition-delay: 200ms;
   }
 `;
