--- frontend-web-expert/src/router/router-map-type.ts
@@ -1,10 +1,0 @@
-import { SvgIconComponent } from "@mui/icons-material";
-import { ComponentType } from "react";
-
-export type RouterType = {
-  path: string;
-  icon: SvgIconComponent;
-  korName: string;
-  engName: string;
-  element: ComponentType;
-};
