+++ frontend-web-expert/src/util/slient-refresh.ts
@@ -1,5 +1,6 @@
 import { client } from "@/api/aixos";
 import { reissue } from "@/api/auth";
+import { LOGIN_ROUTE, MAIN_DASHBOARD_ROUTE } from "@/router/router";
 import { redirect } from "react-router-dom";
 
 export const TOKEN_EXPIRE_MS = 60 * 60 * 1000;
@@ -13,13 +14,13 @@ export const onSlientRefresh = () => {
       setTimeout(onSlientRefresh, TOKEN_EXPIRE_MS - 60 * 1000);
       localStorage.setItem("logged", "true");
 
-      if (window.location.pathname.includes("/login")) {
-        redirect("/");
+      if (window.location.pathname.includes(LOGIN_ROUTE)) {
+        redirect(MAIN_DASHBOARD_ROUTE);
       }
     })
     .catch((error) => {
       console.log(error);
       localStorage.setItem("logged", "false");
-      redirect("/login");
+      redirect(LOGIN_ROUTE);
     });
 };
